creation d un twig _form pour eviter de repeter le code

+++ src/Controller/PinsController.php
@@ -99,12 +99,17 @@ class PinsController extends AbstractController
     }
 
     /**
-     * @Route("/pins/{id<[0-9]+>}/delete", name="pins_delete", methods={"GET", "DELETE"})
+     * @Route("/pins/{id<[0-9]+>}/delete", name="pins_delete", methods={"DELETE"})
      */
-    public function delete(EntityManagerInterface $em, Pin $pin): Response
+    public function delete(Request $request, EntityManagerInterface $em, Pin $pin): Response
     {
-        $em->remove($pin);
-        $em->flush();
+        //verificatin du token
+        if ($this->isCsrfTokenValid('pin_deletion' . $pin->getId(),  $request->request->get('csrf_token') ))
+        {
+            $em->remove($pin);
+            $em->flush();
+        }
+
 
         return  $this->redirectToRoute('home');
     }
